requirements.txt
streamlit>=1.32.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
pandas>=2.0.0
pytesseract>=0.3.10
openai>=1.30.0
